CHANGE LIST .782b
1. Fixed an error in the definitions for the A8DH100P and A2ZJA00P roms. This error caused an unintended
change to specific values in these rom (3 bytes) when the rom was saved. The updated definitions contain
a fix which will return these to the correct factory settings after the rom is saved. The error also
involved the CEL fixes which were not defined correctly for these roms. If you have disabled any CELs
with these roms, be sure to renable them with the old definitions before using these new definitions. Roms
other than the A8DH100P and the A2ZJA00P do NOT have any of these issues.

2. New rom definition added:
- A4TC300K (03 USDM WRx AT)

3. The conversion formula for the various Timing Compensation tables has been changed from a percentage
to actual degrees of correction. The previous formula was not an accurate representation in terms of
the actual percentage of timing correction. [Credit Freon]

git-svn-id: http://svn.3splooges.com/romraider-arch@359 d2e2e1cd-ba16-0410-be16-b7c4453c7c2d